requirements.txt
python-docx>=1.1.0
requests>=2.28.0
beautifulsoup4>=4.12.0
